requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
numpy>=1.24.0
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
scikit-learn>=1.3.0

# CRITICAL: PyTorch and transformers versions - DO NOT CHANGE
# These specific versions are required for semantic skill matching
# torch 2.9.0 has meta tensor bugs, torch 2.5.1 has autocast import errors
torch==2.6.0
sentence-transformers==2.7.0

python-dotenv>=1.0.0
gunicorn>=21.0.0
setuptools>=68.0.0
wheel>=0.40.0
requests>=2.31.0
PyPDF2>=3.0.0
docx2txt>=0.8
pdfminer.six>=20231228
pymongo>=4.8.0
google-generativeai>=0.3.0
razorpay>=2.0.0
moviepy>=1.0.3
openai-whisper>=20231117
deep-translator>=1.11.4
elevenlabs>=0.2.0
yt-dlp>=2023.10.0
# Chatbot API dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
motor>=3.3.0
pydantic>=2.6.0
